requirements.txt
fastapi
uvicorn[standard]
anthropic
ddgs
python-dotenv
motor
pymongo
